chore: Split auth routes into separate router
